Add license file
Standardization in license documentation is important because, in addition to making it easy for humans to find this
vital information, it allows machines to automate the process of license type determination, which is useful both for
discovering suitable open source projects as well as checking open source license compliance.

Previously, the open source license of this library was not documented in a dedicated license file. Due to this, even
though the library is licensed under a standardized open source license, the open license type could not be identified
with 100% confidence by machines (e.g., the Licensee Gem used by the GitHub website), which meant identification could
only be made by a human carefully evaluating the license text.

The project's license is hereby documented in a standardized license file. The result is that the project's open source
license type can now be automatically identified. This does not result in any change to the licensing of the library
source code (which consists solely of files extracted from CMSIS).

The alternative approach would be to configure the library generation script to extract the license file from CMSIS.
However, that would not accomplish the goal of achieving license type recognition by GitHub, and by systems which use
the license type data provided by the GitHub API or equivalent detection approaches.